moved old files to generate-routes and common + authentication throws error instead of returning it
