+++ lib/generate-routes/octane-routes.js
@@ -21,7 +21,7 @@ const fs = require('fs')
 const path = require('path')
 const request = require('request')
 
-const Query = require('./query')
+const Query = require('../query')
 const Reference = require('./models/reference')
 const MultiReference = require('./models/multi-reference')
