+++ lib/root/requestHandler.js
@@ -109,17 +109,10 @@ class RequestHandler {
     }
 
     logger.debug('Signing in.')
-    const error = await this._requestor.post(authOptions).catch(err => {
-      return err
-    })
-
-    if (error) {
-      throw error
-    }
+    await this._requestor.post(authOptions)
 
     logger.debug('Signed in request - Success.')
     this._isAuthenticated = true
-    return this
   }
 
   /**
@@ -129,15 +122,9 @@ class RequestHandler {
   async signOut () {
     logger.debug('Signing out.')
 
-    const error = await this._requestor.post({ url: '/authentication/sign_out' }).then(() => {
+    await this._requestor.post({ url: '/authentication/sign_out' }).then(() => {
       this._isAuthenticated = false
-    }).catch((err) => {
-      return err
     })
-
-    if (error) {
-      throw error
-    }
 
     logger.debug('Signed out.')
     return this
@@ -160,9 +147,7 @@ class RequestHandler {
         this._needsAuthenication = false
       }
 
-      return this._requestor[method](requestOptions).catch(err => {
-        return err
-      })
+      return this._requestor[method](requestOptions)
     })
   }
 }
