+++ lib/root/urlBuilder.js
@@ -21,7 +21,6 @@
  * @param {Number} sharedSpace - Octane shared space id
  * @param {Number} workspace - Octane workspace id
  */
-
 class UrlBuilder {
   constructor (baseUrl, sharedSpace, workspace) {
     this.baseUrl = baseUrl
@@ -55,7 +54,7 @@ class UrlBuilder {
     this.requestUrl.query = `"${query}"`
   }
 
-  script() {
+  script () {
     this.requestUrl.script = true
   }
 
@@ -96,7 +95,7 @@ class UrlBuilder {
     }
     if (this.requestUrl.at) {
       builtUrl = `${builtUrl}/${this.requestUrl.at}`
-      if(this.requestUrl.script && this.requestUrl.entityName === 'tests') {
+      if (this.requestUrl.script && this.requestUrl.entityName === 'tests') {
         builtUrl = `${builtUrl}/script`
       }
       if (this.requestUrl.fields) {
